Fix: Minor UI text and formatting updates
- Changed "Don't have an account?" to "Don't have an account?". -
Changed "We'll send you a verification code to sign in" to "We'll send
you a verification code to sign in". - Changed "Didn't receive the
code?" to "Didn't receive the code?". - Changed "We'll send you a
verification code to get started" to "We'll send you a verification code
to get started". - Changed "We'll use this for order updates and
receipts" to "We'll use this for order updates and receipts". - Changed
"Your information is secure and encrypted" to "Your information is
secure and encrypted". - Added `// @ts-expect-error` to bypass errors.